Cherry pick: Change floats to doubles for path lengths, change a bunch of functionality for tests
Conflicts:
	src/main/AlgoTest.java
	src/main/AnyAnglePathfinding.java
	src/main/Experiment.java

+++ src/main/Visualisation.java
@@ -52,7 +52,7 @@ public class Visualisation {
                 gridLineSet.addLine(path[i][0], path[i][1],
                         path[i+1][0], path[i+1][1], Color.BLUE);
             }
-            float pathLength = Utility.computePathLength(gridGraph, path);
+            double pathLength = Utility.computePathLength(gridGraph, path);
             System.out.println("Path Length: " + pathLength);
             
             boolean isTaut = Utility.isPathTaut(gridGraph, path);
